Refactor "sanity check" into a separate "type check" module.
