+++ src/core/validation/function_definition.rs
@@ -1,7 +1,7 @@
+use crate::core::validation::type_check::CanTypeCheck;
 use crate::core::validation::{
     apply_rule_10301, apply_rule_10307, apply_rule_10308, apply_rule_10309, apply_rule_10310,
-    apply_rule_10312, apply_rule_10401, apply_rule_10402, apply_rule_10404, SanityCheckable,
-    SbmlValidable,
+    apply_rule_10312, apply_rule_10401, apply_rule_10402, apply_rule_10404, SbmlValidable,
 };
 use crate::core::{FunctionDefinition, SBase};
 use crate::xml::{OptionalXmlChild, OptionalXmlProperty, XmlWrapper};
@@ -37,4 +37,4 @@ impl SbmlValidable for FunctionDefinition {
     }
 }
 
-impl SanityCheckable for FunctionDefinition {}
+impl CanTypeCheck for FunctionDefinition {}
