+++ src/core/validation/xml_definitions.rs
